refactor: GTFS Realtime fetching, saving, state update, event detection and publication
