+++ backend/infobus/celery.py
@@ -31,11 +31,11 @@ def debug_task(self):
 app.conf.beat_schedule = {
     "update-gtfs-schedule": {
         "task": "engine.tasks.update_gtfs_schedule",
-        "schedule": crontab(minute=0),
+        "schedule": crontab(minute=30),
     },
     "update-gtfs-realtime": {
         "task": "engine.tasks.update_gtfs_realtime",
-        "schedule": timedelta(seconds=15),
+        "schedule": timedelta(seconds=30),
     },
     "save-gtfs-realtime": {
         "task": "engine.tasks.save_gtfs_realtime",
