+++ backend/pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "requests>=2.33.1",
     "pyarrow>=21.0.0",
     "duckdb>=1.5.3",
+    "pydantic>=2.13.4",
 ]
 
 [dependency-groups]
