+++ backend/runs/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from .models import Run
 
 # Register your models here.
+
+admin.site.register(Run)
