+++ backend/updates/builders/trip/occupancy_status.py
@@ -1,0 +1,11 @@
+"""
+In this file, I need to build the messages that will be sent
+
+{
+    'transit_system': 'mbta',
+    'event_id': '019fc0a8-c557-70ed-a435-cf215e3412ac',
+    'event_type': 'OccupancyStatusChanged',
+    'run_id': '019fc09f-2af5-75b0-baab-c9ec701db592',
+    'current_state': '1'
+}
+"""
