+++ backend/website/templates/index.html
@@ -48,15 +48,6 @@
             >Documentación <strong>Infobús API v1.0</strong></a
           >
         </p>
-        <p>
-          <i class="bi bi-clipboard2-pulse-fill"></i>
-          <a
-            href="{% url 'status' %}"
-            class="text-decoration-none"
-            target="_blank"
-            >Estado del sistema</a
-          >
-        </p>
         <span class="badge rounded-pill text-bg-warning">En desarrollo</span>
       </div>
     </main>
