+++ compose.dev.yml
@@ -121,7 +121,7 @@ services:
       retries: 5
 
   memory:
-    image: redis:7-alpine
+    image: redis:8-alpine
     command: ["redis-server", "--appendonly", "yes"]
     ports:
       - "${MEMORY_PORT:-6379}:6379"
